+++ pybamm/input/parameters/lithium_ion/Marquis2019.py
@@ -424,6 +424,7 @@ def get_parameter_values():
         "Positive current collector thermal conductivity [W.m-1.K-1]": 237.0,
         "Nominal cell capacity [A.h]": 0.680616,
         "Current function [A]": 0.680616,
+        "Contact resistance [Ohm]": 0,
         # negative electrode
         "Negative electrode conductivity [S.m-1]": 100.0,
         "Maximum concentration in negative electrode [mol.m-3]": 24983.2619938437,