+++ pybamm/input/parameters/lithium_ion/NCA_Kim2011.py
@@ -368,6 +368,7 @@ def get_parameter_values():
         "Positive current collector thermal conductivity [W.m-1.K-1]": 158.079,
         "Nominal cell capacity [A.h]": 0.43,
         "Current function [A]": 0.43,
+        "Contact resistance [Ohm]": 0,
         # negative electrode
         "Negative electrode conductivity [S.m-1]": 100.0,
         "Maximum concentration in negative electrode [mol.m-3]": 28700.0,