+++ pybamm/input/parameters/lithium_ion/Ramadass2004.py
@@ -437,6 +437,7 @@ def get_parameter_values():
         "Positive current collector thermal conductivity [W.m-1.K-1]": 237.0,
         "Nominal cell capacity [A.h]": 1.0,
         "Current function [A]": 1.0,
+        "Contact resistance [Ohm]": 0,
         # negative electrode
         "Negative electrode conductivity [S.m-1]": 100.0,
         "Maximum concentration in negative electrode [mol.m-3]": 30555.0,