+++ pybamm/input/parameters/lithium_ion/Xu2019.py
@@ -269,6 +269,7 @@ def get_parameter_values():
         "Electrode width [m]": 0.0154,
         "Nominal cell capacity [A.h]": 0.0024,
         "Current function [A]": 0.0024,
+        "Contact resistance [Ohm]": 0,
         # negative electrode
         "Negative electrode OCP [V]": 0.0,
         "Negative electrode conductivity [S.m-1]": 10776000.0,