+++ pybamm/models/submodels/electrode/base_electrode.py
@@ -96,7 +96,9 @@ class BaseElectrode(pybamm.BaseSubModel):
 
         return variables
 
-    def _get_standard_current_collector_potential_variables(self, phi_s_cn, phi_s_cp):
+    def _get_standard_current_collector_potential_variables(
+        self, phi_s_cn, phi_s_cp, delta_phi_contact
+    ):
         """
         A private function to obtain the standard variables which
         can be derived from the potentials in the current collector.
@@ -107,6 +109,8 @@ class BaseElectrode(pybamm.BaseSubModel):
             The potential in the negative current collector.
         phi_s_cp : :class:`pybamm.Symbol`
             The potential in the positive current collector.
+        delta_phi_contact : :class:`pybamm.Symbol`
+            The potential difference due to the contact resistance, if any.
 
         Returns
         -------
@@ -127,7 +131,8 @@ class BaseElectrode(pybamm.BaseSubModel):
             "Negative current collector potential [V]": phi_s_cn,
             "Positive current collector potential [V]": phi_s_cp,
             "Local voltage [V]": V_cc,
-            "Terminal voltage [V]": V,
+            "Terminal voltage [V]": V - delta_phi_contact,
+            "Contact overpotential [V]": delta_phi_contact,
         }
 
         return variables
@@ -166,9 +171,15 @@ class BaseElectrode(pybamm.BaseSubModel):
             phi_s_cn = variables["Negative current collector potential [V]"]
             phi_s_p = variables["Positive electrode potential [V]"]
             phi_s_cp = pybamm.boundary_value(phi_s_p, "right")
+            if self.options["contact resistance"] == "true":
+                param = self.param
+                I = variables["Current [A]"]
+                delta_phi_contact = I * param.R_contact / param.potential_scale
+            else:
+                delta_phi_contact = pybamm.Scalar(0)
             variables.update(
                 self._get_standard_current_collector_potential_variables(
-                    phi_s_cn, phi_s_cp
+                    phi_s_cn, phi_s_cp, delta_phi_contact
                 )
             )